feat(metrics): extract unified calculation service and eliminate code duplication
- Create MetricsCalculatorService to eliminate duplicate calculation logic
- Remove 13 duplicate functions from metrics API routes
- Reduce code size: metrics routes 46-51% smaller (350+ lines saved)
- Follow Service Layer principle - business logic extracted to services
- Improve maintainability and testability with centralized calculations
- All validation passes: Build ✓ Lint ✓ Typecheck ✓ Tests (24/24 ✓)
