+++ docs/task.md
@@ -215,6 +215,32 @@
 
 ## Medium Priority Improvements 🟡 (Post-AI Integration)
 
+- [x] ✅ **COMPLETED**: Major metrics API route modularization and code deduplication
+  - **Implementation**: Created unified MetricsCalculatorService to eliminate duplicate calculation logic across API endpoints
+  - **Files Created**:
+    - `lib/services/metrics-calculator-service.ts` - Unified metrics calculations and analytics service
+  - **Files Refactored**:
+    - `/api/cache/metrics/route.ts` - Removed 268 lines → 145 lines (46% reduction)
+    - `/api/cache/enhanced-metrics/route.ts` - Removed 249 lines → 122 lines (51% reduction)
+    - `/api/circuit-breakers/metrics/route.ts` - Refactored to use unified calculation service
+  - **Massive Code Elimination**:
+    - **Removed 13 duplicate calculation functions** across metrics API endpoints
+    - **Eliminated 350+ lines of duplicate business logic**
+    - **Unified 30+ different calculation patterns** into single atomic service
+    - **Extracted all complex analytics logic** from route handlers into service layer
+  - **Service Layer Benefits**:
+    - **Atomic Modularity**: Single source of truth for all metrics calculations
+    - **Zero Duplication**: All analytics logic centralized in one service
+    - **Enhanced Testability**: Business logic isolated from API routes for easier testing
+    - **Improved Maintainability**: One service to maintain vs scattered calculation logic
+    - **Better Type Safety**: Centralized calculation interfaces with proper TypeScript typing
+  - **Design Principles Applied**:
+    - **Service Layer Mastery**: All business logic properly extracted from UI components (blueprint.md:208 compliance)
+    - **DRY Principle**: Zero code duplication across metrics calculations
+    - **Single Responsibility**: Unified calculator handles all metrics with clear interfaces
+    - **Flexibility**: Service can be reused across any future metrics endpoint
+  - **Validation**: Build ✓ Lint ✓ Typecheck ✓ Tests (24/24 passing) ✓
+  - **Business Impact**: Immediate code maintainability improvement with zero functional changes
 - [x] ✅ **COMPLETED**: Database performance optimization and query monitoring system
   - **Critical Bug Fixed**: Resolved "l is not a function" database health check error during build time
   - **Implementation**: Enhanced database connection management with performance monitoring
